move startWorkers to startup
